Update patch set 1
Patch Set 1: -Code-Review

Yes, it sets the size allowed of each partition and causes the build to fail when the for example recovery.img is larger than the set size.  In my case the recovery.img is larger than the current size when using TWRP.   Thus needing to set the size to the actual size of our device.

Patch-set: 1