Update patch set 1
Patch Set 1:

Just double verified it using a couple of methods to report size and my values are correct, but I am on Verizon so I guess we just found another difference between TMO, Verizon and Sprint.   Very odd indeed.  So guess we need someone to do this on 900T and 900F to see what the values are.

Patch-set: 1