Update patch set 1
Patch Set 1:

ok I need input on this. This only works if we swap to the KK HLTE libsec-ril.so and libqmiservies.so I know its not ideal but as the vendor libs stand now data wont even work nor phone calls without this patch. My guess is our device has that stupid feature where you can share WiFi and mobile data to get a faster internet connection. I'm thinking there is some stub in there that is like either 1 or 0 of its on or off and we need to set a dummy function to always be 0 but I don't know where to look.

Patch-set: 1