Update patch set 1
Patch Set 1:

no. the only 4.4.4 release there is a a SPR leak. and the tar wont extract for me. its curupted somehow. Garwynn flashed it directly to his sprint model. I was gonna update all libs to 4.4.4 as well and fingerprints but hit a road block when the leak wouldnt extract. So yes these are the newest official fingerprints. they have stuck with 4.4.2 it seems

Patch-set: 1