Update patch set 3
Patch Set 3:

Yup, 14.0. Most of the other devices are using a property that they can do an equivalence comparison to, but we get a bunch of versioning after the variant. Something like "G900TUVS1GPG2" will never == "G900T", so everything would end up falling down to the default case. 

find() will get us what strstr does and comparing ==0, as opposed to !=std::string::npos, is probably most correct because we want to look at the start of the string.

Maybe there's another readonly unique value we can use, but this is probably good enough.

Patch-set: 3